backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
PyPDF2==3.0.1
pdfminer.six==20221105
sentence-transformers==2.2.2
torch==2.1.0
transformers==4.35.2
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
pydantic==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
aiofiles==23.2.1 